requirements.txt
flask
packaging
requests
mcstatus
python-dotenv
